Modificaciones proyecto store

+++ Kevin Martin Del Campo/desarolla.Store/routers/users.js
@@ -17,7 +17,7 @@ const Utils = require('../utils/utils');
 router.get('/all', async (req, res) => {
 
     var userIsAdmin = await Utils.isAdmin(req, res);
-    if(!userIsAdmin) {
+    if (!userIsAdmin) {
         return;
     }
 
@@ -75,6 +75,7 @@ router.post('/register', async (req, res) => {
             email: datosUsuario.email
         }]
     });
+
     if (userExists) {
         return res.status(401).send({
             error: "El usuario con este nickname/correo ya existe"
@@ -99,10 +100,11 @@ router.put('/:nickname', async (req, res) => {
     const nickname = req.params.nickname;
     const usuarioActual = req.cookies["SESSIONID"];
     const userData = req.body;
+    var userIsAdmin = false;
 
-    if(nickname !== usuarioActual) {
-        var userIsAdmin = await Utils.isAdmin(req, res);
-        if(!userIsAdmin) {
+    if (nickname !== usuarioActual) {
+        userIsAdmin = await Utils.isAdmin(req, res);
+        if (!userIsAdmin) {
             return;
         }
     }
@@ -137,6 +139,38 @@ router.put('/:nickname', async (req, res) => {
                 user.name = userData.name
                 break;
 
+            case "nickname":
+                var newNickname = userData.nickname;
+                var userExists = await User.findOne({
+                    nickname: newNickname
+                });
+
+                if (userExists) {
+                    res.status(403).send({
+                        error: "El nickname: " + newNickname + " ya está ligado a otro usuario"
+                    });
+                    return;
+                }
+
+                user.nickname = newNickname;
+                break;
+
+            case "email":
+                var newEmail = userData.email;
+                var userExists = await User.findOne({
+                    email: newEmail
+                });
+
+                if (userExists) {
+                    res.status(403).send({
+                        error: "El email: " + newEmail + " ya está ligado a otro usuario"
+                    });
+                    return;
+                }
+
+                user.email = newEmail;
+                break;
+
             case "lastName":
                 user.lastName = userData.lastName
                 break;
@@ -152,6 +186,12 @@ router.put('/:nickname', async (req, res) => {
             case "address":
                 user.address = userData.address
                 break;
+
+            case "userType":
+                if (userIsAdmin) {
+                    user.userType = userData.userType;
+                }
+                break;
         }
     }
 
@@ -166,16 +206,31 @@ router.put('/:nickname', async (req, res) => {
 router.delete('/:nickname', async (req, res) => {
 
     var userIsAdmin = await Utils.isAdmin(req, res);
-    if(!userIsAdmin) {
+    if (!userIsAdmin) {
         return;
     }
 
     var parametros = req.params;
     var nickname = parametros.nickname;
 
+    /*var user = await User.findOne({ nickname: nickname });
+    if(!user) {
+        res.send({
+            message: "El usuario: " + nickname + " no existe"
+        });
+        return;
+    }*/
+
     var usuarioBorrado = await User.deleteOne({
         nickname: nickname
     });
+
+    if(usuarioBorrado.deletedCount === 0) {
+        res.status(404).send({
+            message: "El usuario: " + nickname + " no existe"
+        });
+        return;
+    }
 
     res.send({
         message: "Se ha borrado el usuario: " + nickname
@@ -212,7 +267,7 @@ router.post('/login', async (req, res) => {
         password: datosLogin.password
     });
 
-    if(!usuario) {
+    if (!usuario) {
         return res.status(404).send({
             error: "Datos incorrectos de inicio de sesión. Verifique el user/password"
         });
